More work on implementing maps.
They type I really want to use for Maps is DictionaryEntry, but that's a
struct so we won't treat it properly (we don't instrument properties
right now since we don't know that they're pure).

For now I made my own non-generic tuple class. Eventually we should be
able to investigate the fields of the key/value's type but that can't
be done with just the non-generic tuple as is.