Fix issue 76 -- multiple exceptional exits when Thread.Abort() is called. Celeriac now produces Daikon-vald output for the finally-guard test case.
I fixed this by adding a test to see if the exception caught by our runtime
exception handler is a thread abort exception, if it is then we don't instrument
it. This could lead to some unmatched returns when a thread is aborted and there
is no normal exit from any exception handler. However this I think mirrors the
actual behavior -- the method was terminated without reaching any exit. The
ThreadAbort test function could be expanded to still instrument the function
if it has never been visited before, and the other exception printing functions
would keep a list of visited exceptions.